refactor: move remaining network-dependent tests to integration
Move tests that call HuggingFace Hub APIs out of tests/unit/:

- test_text_classification.py → tests/integration/datasets/
  (every test downloads GLUE/MRPC via load_dataset)

- TestResolveTaskAndModelClass and TestConflictScenarios from
  test_hf_model_class_mapping.py → tests/integration/loader/
  (13 tests call AutoConfig.from_pretrained, missed in #95)

Fixes #253
